Move boilerplate into component directory.
