+++ myuw/static/vue/components/_templates/boilerplate/search.vue
@@ -1,0 +1,13 @@
+<template>
+  <div>
+    this is search
+  </div>
+</template>
+
+<script>
+export default {
+  data: function() {
+    return {};
+  },
+};
+</script>
