+++ myuw/static/vue/home.js
@@ -20,7 +20,7 @@ import {
 } from 'bootstrap-vue';
 
 // layout componenets
-import Boilerplate from './components/_templates/boilerplate.vue';
+import Boilerplate from './components/_templates/boilerplate/boilerplate.vue';
 
 // common components
 import Grades from './components/_common/grades.vue';
